refactor: move OS theme detection to App.tsx
