+++ web/src/App.tsx
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
+import { useEffect } from "react";
 import { RouterProvider } from "@tanstack/react-router"
 import { QueryClientProvider } from "@tanstack/react-query";
 import { Toaster } from "react-hot-toast";
@@ -10,7 +11,7 @@ import { Portal } from "react-portal";
 import { Router } from "@app/routes";
 import { routerBasePath } from "@utils";
 import { queryClient } from "@api/QueryClient";
-import { AuthContext } from "@utils/Context";
+import { AuthContext, SettingsContext } from "@utils/Context";
 
 declare module '@tanstack/react-router' {
   interface Register {
@@ -19,6 +20,18 @@ declare module '@tanstack/react-router' {
 }
 
 export function App() {
+  const [_, setSettings] = SettingsContext.use();
+
+  useEffect(() => {
+    const themeMediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
+    const handleThemeChange = (e: MediaQueryListEvent) => {
+      setSettings(prevState => ({ ...prevState, darkTheme: e.matches }));
+    };
+
+    themeMediaQuery.addEventListener('change', handleThemeChange);
+    return () => themeMediaQuery.removeEventListener('change', handleThemeChange);
+  }, [setSettings]);
+
   return (
       <QueryClientProvider client={queryClient}>
         <Portal>
